docs: Add quickstart guide for local QEMU/KVM VMs
Commit e2b58d60 ("tools: Add start-local-vm script") added a wrapper
script to simplify running a locally built Bottlerocket image in a local
VM using QEMU and KVM. Provide a quickstart guide analogous to the
existing ones for EKS, ECS, and VMware to help make use of it.

The guide currently assumes anyone following along to be running Fedora.
Changes to the guide for other distros will be minimal, if they are
needed at all. However, changes in start-local-vm are likely to be
required. Leave this to a dedicated later change.

Signed-off-by: Markus Boehme <markubo@amazon.com>