kernel: config: clean up '=n' settings
Disabling a kernel config setting is usually done in a config file in
the form of

\# CONFIG_<CONFIG_NAME> is not set

instead of

CONFIG_<CONFIG_NAME>=n

While both today should be equivalent historically there was problems
with specifying '=n' still making that setting defined and having the
opposite effect of what is intended by setting '=n'. Nowadays either way
will result in a final config that will only have the `is not set`
variant for all deactivated options, so the diff of the resulting
artifact before and after this patch will be empty.

As the `is not used` form is the commonly used form, lets adhere to it.
As a bonus it reduces the noise when comparing configs that are manually
merged using the merge-config script.

Replaced all the '=n' settings by the common form through command:

for file in packages/kernel*/config*;
do
    sed -E -i "s/^(CONFIG_[A-Z0-9_]*)=n/# \1 is not set/g" $file;
done

Signed-off-by: Leonard Foerster <foersleo@amazon.com>