Format issue field in PR template
We have had several instances where the issue associated with a PR has
not been automatically closed. Engineers then need to know to track down
the issue after the merge to manually close out the issue.

Part of this appears to be due to the PR not having the format expected
to properly link the PR to the Issue [0]. This updates our pull request
template to format this placeholder string to help guide the user
towards using the necessary format to link and close the issue.

[0] https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword

Signed-off-by: Sean McGinnis <stmcg@amazon.com>