kernel: net: Add AMD 10G+ network drivers
We are interested in AMD_XGBE. Explicitly disable DCE for that
driver as we likely do not need DCB.

In addition this driver will select option AMD_XGBE_HAVE_ECC on X86
platforms. As this config fragment is used on both x86 and aarch64
platforms we can not explicitly set it.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>