testsys: Enable multiple tests on a cluster
TestSys uses labels on the crd's to determine which ec2 resources rely
on a specific cluster. If a new instance provider requires a cluster that
is already created, TestSys marks the new instance provider as conflicting with
all other ones which prevents the new instances from being created until the
other instance providers have been deleted.