buildsys: add publish image size hint to manifests
Allow manifests to specify a hint for the published image sizes, for
the common case where additional storage space is desired on the data
volume.

A helper function is provided to calculate the sizes for the OS and
data volumes, based on the image layout and the publish size hint.

Signed-off-by: Ben Cressey <bcressey@amazon.com>