kernel: ignore initramfs from bootloader
For Secure Boot support, the GRUB configuration (grub.cfg) will need
to be signed by a trusted key, so that the chain of trust extends to
the kernel command line, which contains sensitive data such as the
dm-verity root hash.

However, Bottlerocket also allows the boot config feature to be used
to extend the kernel command line. Boot config data is passed by the
bootloader using the "initrd" mechanism; it consists of a payload at
the end of the initramfs cpio buffer list.

In normal use, the "initrd" file should only ever contain boot config
data. However, because this file is locally generated, it cannot be
signed by a key known to the chain of trust. Because it is treated as
an initramfs, it could be leveraged to run an unverified userspace by
prepending it with an actual initramfs.

This change causes the kernel to discard any initramfs or initrd data
passed by the bootloader. Any boot config data remaining at the end
of the input will still be processed, as intended.

Without Secure Boot support, ignoring the initramfs is not useful as
a security measure, but it still improves robustness by preventing
accidental usage that might lead to boot failures.

Signed-off-by: Ben Cressey <bcressey@amazon.com>