build: set target CPU to target arch for RPMs
This causes the RPM package architecture to match the target's rather
than the host's. By default, `dnf` and `rpm` will ignore non-native
RPMs, but both can be forced to use them.

When the package architecture is set correctly, it's not necessary to
include the architecture in the package name, and the macros in the
SDK no longer do so. Fix up various places in the image build process
that assumed the architecture would be present.

Signed-off-by: Ben Cressey <bcressey@amazon.com>