release: special-case data partition handling
We reconceptualize DATA-A and DATA-B as a set of preferred and fallback
data partitions.

The preferred partition will always get repart'd if the partition
matching the preferred data partition GUID exists.

The fallback partition only gets repart'd if we've confirmed that the
data partition resides on a device that is not EBS or Xen virtual disk.
We determine this by checking new device symlinks that indicates the
storage device type. This avoids any behavior change for EC2 instance
launches, even if additional size is allocated to the first EBS volume.

This is done to support platforms like Snow, where the same AMI is used
but a second volume is not available at launch. Since Snow does not
handle EBS volumes in the same way, the symlink checks do not prevent
the fallback partition on the first volume from being used.