start-local-vm: add options to override firmware
Both new options are meant to help with Secure Boot testing.

The `firmware-code` option allows a custom edk2 build to be used, for
example with aarch64 where the AAVMF build packaged by Fedora doesn't
have the Secure Boot feature available.

The `firmware-vars` option populates the initial firmware variables
from a different file, and sets it to the correct size for aarch64.

Signed-off-by: Ben Cressey <bcressey@amazon.com>