build: auto-create a local Secure Boot profile
Ensure that at least one Secure Boot profile is always available, for
cases where the variant has Secure Boot enabled.

This uses the "local" version of the script, since locally generated
keys cost nothing and are guaranteed to be available. This is similar
to the locally generated keys used by default for TUF repositories,
in that neither is suitable for long-term production use.

The individual Secure Boot profile and the directory where profiles
are stored can both be overridden, so that profiles can be stored in
a different location, such as another Git repository.

Profiles are checked for completeness, to allow the expected files to
evolve over time, for example to add support for a new platform that
expects EFI variables in a different format.

Signed-off-by: Ben Cressey <bcressey@amazon.com>