Dynamically get list of variants for GitHub Actions
This updates the `build` workflow to determine the set of variants to
test against at runtime. It looks at what is under the `variants`
directory to determine what to include, rather than needing to hard code
this list in the action itself.

This also gives the potential that we can use this in a composable
workflow so multiple workflows can get the list of variants without
needing to remember to update the list in multiple places whenever a
variant is added or removed.