kernel-6.1: increase default net.unix.max_dgram_qlen to 512
[port from kernel-5.15: ddc075761b315d76f86668a9552bc524591aaf5e]

Increase the kernel's default value for the net.unix.max_dgram_qlen sysctl
to 512. This is a change to the kernel rather than a plain sysctl setting
since systemd-sysctl only applies settings to the host, while the changed
default value in the kernel also applies to every new network namespace.

Signed-off-by: Markus Boehme <markubo@amazon.com>

Signed-off-by: Leonard Foerster <foersleo@amazon.com>