build: retry builds for createrepo_c error
RPMs from package builds go into a shared host directory which is
mounted into subsequent package builds. It is possible for new RPMs
to be copied in just as `createrepo_c` is running, which can cause it
to read partial RPMs and return an error.

Since the partial RPMs can't be required dependencies - or else the
build would not have started yet - it would be possible to ignore the
`createrepo_c` errors and continue with whatever repo was created.
However, this could mask other errors that might not be expected.

`buildsys` copies RPMs into a different directory on the filesystem
before renaming them into the shared directory. Calling `sync()` on
one or both directories might help, but would hurt the performance of
all builds for what seems to be a rare error.

Instead, treat the specific `rpmReadPackageFile()` error as retryable
and attempt the build again. If a genuinely malformed RPM is written
to the shared directory, the builds will ultimately fail when retries
are exhausted.

Signed-off-by: Ben Cressey <bcressey@amazon.com>