kernel-5.10: Choose upstream smartpqi driver over backport
In the past we have built-in the upstream smartpqi driver for our metal
platforms and inherited the module build on aws/vmware platforms from
AL. AL recently introduced a newer version of that driver alongside the
in-tree driver. Keep the status quo by only building the in-tree
variant, effectively reducing the amount of code built and shipped.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>