tools/diff-kernel-config: adapt to new RPM naming scheme
With the recent cut-over to the RPM macros being sourced from the SDK
(commit 6aed1b97f4c5) the naming scheme for RPMs changed. The target
architecture is no longer custom encoded in the prefix of the name (with
the architecture suffix indicating which architecture the package has
been built on), but properly reflected in the suffix.

Adapt the `diff-kernel-config` to deal with that by adding another glob
pattern. Resolve ambiguity if a change is diffed before and after that
name change by picking the latest build. For determining the kernel
version, go straight to the RPM meta data instead of trying to pick
apart the name--due to `rpm2cpio` we already have a dependency on the
`rpm` package anyway.

Signed-off-by: Markus Boehme <markubo@amazon.com>