actions: Use larger runners for nightly checks
The nightly checks were recently added for performing extra validation
by bypassing the GOPROXY cache. These have been failing due to using the
default runners for execution.

This updates the job definition to use the same runners we use for the
per-PR build workflow to make sure there is enough disk space and other
resources to actually perform the build.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>