build: use static files instead of /dev/null
An unintended consequence of this `path = "/dev/null"` in the Cargo.toml
files of packages and variants is that a change to the device causes a
complete rebuild of all of Bottlerocket. So, for example, restarting a
machine would cause Bottlerocket to rebuild all packages.

This was an issue in the new Twoliter build system because each time
Twoliter is executed, a container is created with a different /dev/null
device, triggering a full rebuild.

Though less cool, this commit points all packages at a single
packages.rs file and all variants at a variants.rs file to fix this
problem.