linux-firmware: Remove a lot of firmware we do not need in Bottlerocket
Bottlerocket is an OS with a comparably cut-down kernel config and as
such does not provide the wide range of drivers one can find in general
purpose linux distributions. Many of these devices, Bottlerocket does
not support have firmware binaries shipped through linux-firmware.

Drop those firmware blobs we do not ship drivers for in Bottlerocket. I
have tried to section the removal into groups of drivers with
commonalities (e.g. Wifi and Bluetooth devices). Each patch has a list
mapping each driver to a specific kernel config option so we can easily
find which firmware binaries to add back into Bottlerocket if we add
drivers.

The patches only touch the WHENCE file and not the actual binaries as
rpm does not like to apply git binary patches. The WHENCE file is a
record of all the firmware binaries in the package with additional
information for each file. On installation the copy-firmware.sh script
uses the information in the file to determine which files to copy, so it
is enough to remove information from WHENCE in order to remove binaries
from the installed directories.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>