tools/diff-kernel-config: Add release number to full kernel version
When saving the full kernel version for later reference in the kernel
mapping we may come across situations where we update to a newer version
of our base kernel srpm from Amazon Linux which added only added patches
from Amazon Linux but stayed at the same upstream base version. The
before and after would be indistinguishable when only comparing the
version. Add the release number to disambiguate in that situation.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>